+++ src/find.ts
@@ -1,4 +1,4 @@
-import {Node} from './nodes';
+import {Node, Parent} from './nodes';
 import {Test, test} from './test';
 import {CONTINUE, EXIT, visit} from './visit';
 
@@ -9,7 +9,7 @@ import {CONTINUE, EXIT, visit} from './visit';
  * @param [tst] - 'is-compatible' test (such as a type).
  */
 export function find<T extends Node>(tree: T | T[], tst: Test<T>) {
-	let is = test(<Test<T>>tst);
+	let is = test(tst);
 	let result: T = undefined;
 	visit(tree, is, function (node) {
 		result = node;
@@ -26,7 +26,7 @@ export function find<T extends Node>(tree: T | T[], tst: Test<T>) {
  * @returns an Array of zero or more Nodes in the tree that matched the condition.
  */
 export function findAll<T extends Node>(tree: T | T[], tst: Test<T>) {
-	let is = test(<Test<T>>tst);
+	let is = test(tst);
 	let results: T[] = [];
 	visit(tree, is, function (node) {
 		results.push(node);
@@ -44,14 +44,14 @@ export function findAll<T extends Node>(tree: T | T[], tst: Test<T>) {
  *          The first element in the "path" array will be the ancestor that matched the condition, and the node's immediate parent will be the last element of the array.
  *          If no ancestors were found (or none passed the test), an empty array will be returned.
  */
-export function findAncestor<T extends Node>(node: T, tst: Test<T>) {
-	let is = test(<Test<T>>tst);
+export function findAncestor<T extends Parent>(node: Node, tst: Test<T>) {
+	let is = test(tst);
 	let results: T[] = [];
 	while (node && node.parent) {
-		results.unshift(<any>node.parent);
-		if (is(node.parent))
-			break;
-		node = <any>node.parent;
+		results.unshift(<T>node.parent);
+		if (is(<T>node.parent))
+			return results;
+		node = <T>node.parent;
 	}
-	return results;
+	return [];
 }