+++ Changelog.md
@@ -8,3 +8,7 @@
 ## 1.0.2 / 2020/06/05  
 * Merge the correct local branch.
 * Update ReadMe.
+
+## 1.0.3 / 2020/06/08  
+* Unit tests for, and fixes to, the find module.
+* Declaration fixes for the test module and visit module.